# openapi.yaml
openapi: 3.0.3
info:
  title: Money Flow API
  version: 1.0.0
servers:
  - url: /api
security:
  - ApiKeyAuth: []
paths:
  /health:
    get:
      summary: Health check
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                required:
                  - status
  /categories:
    get:
      summary: List categories
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategory'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
  /categories/{id}:
    put:
      summary: Update category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategory'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
    delete:
      summary: Delete category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
  /transactions:
    get:
      summary: List transactions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
    post:
      summary: Create transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransaction'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
  /transactions/{id}:
    put:
      summary: Update transaction
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTransaction'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
    delete:
      summary: Delete transaction
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
  /recurring-expenses:
    get:
      summary: List recurring expenses for a month
      parameters:
        - in: query
          name: month
          required: false
          schema:
            type: string
            pattern: '^\\d{4}-\\d{2}$'
          description: Month in YYYY-MM format. Defaults to current month in user timezone.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecurringExpenseForMonth'
    post:
      summary: Create recurring expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecurringExpense'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringExpense'
  /recurring-expenses/{id}:
    delete:
      summary: Delete recurring expense
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
  /settings/timezone:
    get:
      summary: Get user timezone (UTC offset)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimezoneSettings'
    put:
      summary: Update user timezone (UTC offset)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimezoneSettings'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimezoneSettings'
  /export/xlsx:
    get:
      summary: Export data to XLSX
      responses:
        '200':
          description: XLSX file
          headers:
            Content-Disposition:
              description: Attachment filename
              schema:
                type: string
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      required:
        - id
        - name
    CreateCategory:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    UpdateCategory:
      type: object
      properties:
        name:
          type: string
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        source:
          type: string
        idempotencyKey:
          type: string
          nullable: true
        categoryId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        date:
          type: string
          format: date
        description:
          type: string
          nullable: true
      required:
        - id
        - categoryId
        - amount
        - date
    CreateTransaction:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        date:
          type: string
          format: date
        description:
          type: string
      required:
        - categoryId
        - amount
        - date
    UpdateTransaction:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        date:
          type: string
          format: date
        description:
          type: string
    RecurringExpense:
      type: object
      properties:
        id:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        dayOfMonth:
          type: integer
          format: int32
        date:
          type: string
          format: date
        description:
          type: string
          nullable: true
      required:
        - id
        - categoryId
        - amount
        - dayOfMonth
        - date
    CreateRecurringExpense:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        dayOfMonth:
          type: integer
          format: int32
        date:
          type: string
          format: date
        description:
          type: string
      required:
        - categoryId
        - amount
        - dayOfMonth
        - date
    RecurringExpenseForMonth:
      type: object
      properties:
        id:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        dayOfMonth:
          type: integer
          format: int32
        date:
          type: string
          format: date
        description:
          type: string
          nullable: true
        scheduledDate:
          type: string
          format: date
        committed:
          type: boolean
      required:
        - id
        - categoryId
        - amount
        - dayOfMonth
        - date
        - scheduledDate
        - committed
    TimezoneSettings:
      type: object
      properties:
        utcOffset:
          type: string
          pattern: '^[+-]\\d{2}:\\d{2}$'
      required:
        - utcOffset