# src/recurring-expenses/recurring-expenses.controller.ts
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  Param,
  Post,
  Query,
  UseGuards,
} from '@nestjs/common';
import { CreateRecurringExpenseDto } from './dto/create-recurring-expense.dto';
import { ListRecurringExpensesQueryDto } from './dto/list-recurring-expenses.dto';
import { RecurringExpenseForMonth, RecurringExpensesService } from './recurring-expenses.service';
import { RecurringExpense } from './recurring-expense.entity';
import { AppApiKeyGuard } from '../auth/app-api-key.guard';

@Controller('recurring-expenses')
@UseGuards(AppApiKeyGuard)
export class RecurringExpensesController {
  constructor(private readonly recurringExpensesService: RecurringExpensesService) {}

  @Get()
  async listForMonth(
    @Query() query: ListRecurringExpensesQueryDto,
  ): Promise<RecurringExpenseForMonth[]> {
    return this.recurringExpensesService.listForMonth(query.month);
  }

  @Post()
  async create(@Body() payload: CreateRecurringExpenseDto): Promise<RecurringExpense> {
    return this.recurringExpensesService.create(payload);
  }

  @Delete(':id')
  @HttpCode(204)
  async remove(@Param('id') id: string): Promise<void> {
    return this.recurringExpensesService.remove(id);
  }
}